OK, listen.
I know this commit is a lot. But sometimes life is a lot.

I don't remember what all I did. I just kept not committing because
I'm commitment-phobic when I don't have a running list of changes
in my head. but sometimes, you just have to do the thing. You have to
commit the code. The code works. This is just where we are. Maybe
not all of the code I'm committing right now is fantastic, maybe some of
it isn't in use because it is broken, but that's a problem for
future-me. current-me is just proud she forced herself to commit her damn
code. finally.