heck yeah uploading files over openwebui
all of it is shoved into some ugly ass babys-first-prompt-format.
i stg these prompt format exist for people that lack the imagination
to do better. but i digress.

this special prompt is passed a system message. so i had to add the
ability to parse system messages to instruct models, and i'm just
passing it as user input but wrapped in <<SYS>>.