feat: add job creation instructions to review_log endpoint
This commit adds instructions on how to create a job for the `review_log` endpoint.

The instructions provide an example `httpx` request that can be used to create a job that triggers the `review_log` endpoint on an interval. This will help users easily set up automated review of conversation logs.